cover_screen/service/pyproject.toml
[project]
name = "cover-screen-service"
version = "0.1.0"
description = "Cover Screen Frame Buffer Service"
readme = "README.md"
requires-python = ">=3.12"

[dependency-groups]
dev = [
  "ruff",
]
raspi = [
  "numpy",
  "pyzmq",
  "gpiozero==2.0.1",
  "lgpio==0.2.2.0",
  "luma.lcd==2.11.0",
]
desktop = [
  "numpy",
  "pyzmq",
  "luma.emulator==1.5.0",
]

[tool.uv]
default-groups = ["dev", "raspi"]
